9irc: note that long PRIVMSGs should be sent in pieces.
Currently if a PRIVMSG is too long then 9irc merely
receives an error message about it on the Errors
channel. The message still appears on the chat win
as though it was sent when, in fact, it wasn't.  Instead,
we should take long PRIVMSGs and break them into
multiple messages that are sent individually.