kdtree: change the benchmarks so that InRange lookups are for a smaller radius.
This showcases the benefit of the K-D tree; before the
radius often contained most of all of the points, and the
simplicity of the linear scan trumped the overhead of the
tree. If the radius is smaller (say 1/10th of the size of
the unit square) the K-D tree has a huge advantage.