kdtree: fix a bug in splitMed.
The previous implementation assumed that everything with an index smaller than the median would fall on the left and everything with an index greater than the median would fall on the right. This is true if nothing is equal, but if the points are integral, it's often the case that there are many ties with the median. This caused index out of bound errors because we incorrectly computed the sizes of the left and right sides of the split.

Fixes issue 18.